+++ source/renderer/app/components/wallet/paper-wallet-certificate/CompletionDialog.js
@@ -6,11 +6,12 @@ import classnames from 'classnames';
 import { defineMessages, intlShape } from 'react-intl';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import SVGInline from 'react-svg-inline';
+import { Link } from 'react-polymorph/lib/components/Link';
+import { LinkSkin } from 'react-polymorph/lib/skins/simple/LinkSkin';
 import Dialog from '../../widgets/Dialog';
 import { getNetworkExplorerUrl } from '../../../utils/network';
 import styles from './CompletionDialog.scss';
 import iconCopy from '../../../assets/images/clipboard-ic.inline.svg';
-import externalLinkIcon from '../../../assets/images/link-ic.inline.svg';
 import InlineNotification from '../../notifications/InlineNotification';
 import { DEVELOPMENT } from '../../../../../common/types/environment.types';
 
@@ -166,15 +167,12 @@ export default class CompletionDialog extends Component<Props, State> {
             </p>
 
             <div className={styles.infoBox}>
-              <span
+              <Link
                 className={styles.link}
                 onClick={() => onOpenExternalLink(cardanoExplorerLink)}
-                role="link"
-                aria-hidden
-              >
-                {cardanoExplorerLink}
-                <SVGInline svg={externalLinkIcon} />
-              </span>
+                label={cardanoExplorerLink}
+                skin={LinkSkin}
+              />
             </div>
           </div>